Inventory and Shop fixes
The issue of an access exception was fixed when adding an item to the player inventory. The solution was I forgot to invoke the constructor for the player inventory. Succsseful tests were conducted on adding, removing, and displaying items from the player inventory. Items will not be added if the max size is reached. Adjusted the max inventory size to be 10. When the shop loads, it now displays the player's inventory at the bottom after the shop inventory.